+++ lib/src/job/launcher/step_launcher.dart
@@ -3,9 +3,9 @@
 // modification, are permitted provided the conditions.
 
 // Project imports:
-import 'package:batch/src/job/context/execution_context.dart';
-import 'package:batch/src/job/event/step.dart';
+import 'package:batch/batch.dart';
 import 'package:batch/src/job/launcher/launcher.dart';
+import 'package:batch/src/job/launcher/parallel_launcher.dart';
 import 'package:batch/src/job/launcher/task_launcher.dart';
 
 class StepLauncher extends Launcher<Step> {
@@ -27,13 +27,17 @@ class StepLauncher extends Launcher<Step> {
       await super.executeRecursively(
         event: step,
         execute: (step) async {
-          await TaskLauncher(
-            context: context,
-            tasks: step.tasks,
-          ).run();
-
-          // Removes step scope parameters set within the step executed last time.
-          super.context.stepParameters.removeAll();
+          if (step.task is Parallel) {
+            await ParallelLauncher(
+              context: context,
+              parallel: step.task,
+            ).run();
+          } else {
+            await TaskLauncher(
+              context: context,
+              task: step.task,
+            ).run();
+          }
         },
       );
     }